Extract protocols into a separated package
This commit just continues extracting independent parts of the client code into separated packages. Now it is turn of the client protocols (messages, call and handshake protocols).
